separated Anchor, Edge, Node from store.ts
