+++ src/lib/models/Edge.ts
@@ -1,0 +1,34 @@
+import type {
+  NodeType,
+  EdgeType,
+  AnchorType,
+  StoreType,
+  ResizeNodeType,
+} from './types';
+import { writable, derived, get, readable } from 'svelte/store';
+import { getNodes, getAnchors, findStore } from '../controllers/storeApi';
+import { stores } from './store';
+
+export class Edge implements EdgeType {
+  constructor(
+    public id: string,
+    public sourceX: number,
+    public sourceY: number,
+    public targetX: number,
+    public targetY: number,
+    public canvasId: string,
+    public label?: string,
+    public type?: string,
+    public labelBgColor?: string,
+    public labelTextColor?: string,
+    public edgeColor?: string,
+    public animate?: boolean,
+    public noHandle?: boolean,
+    public arrow?: boolean
+  ) {}
+
+  // TODO: implement me
+  handleDelete() {
+    console.log('deleting edge not implemented yet');
+  }
+}
